修复E2E测试

+++ northstar-main/src/main/java/org/dromara/northstar/data/jdbc/PlaybackRuntimeRepoAdapter.java
@@ -1,5 +1,7 @@
 package org.dromara.northstar.data.jdbc;
 
+import java.util.Objects;
+
 import org.dromara.northstar.common.model.PlaybackRuntimeDescription;
 import org.dromara.northstar.data.IPlaybackRuntimeRepository;
 import org.dromara.northstar.data.jdbc.entity.PlaybackRuntimeDescriptionDO;
@@ -19,12 +21,18 @@ public class PlaybackRuntimeRepoAdapter implements IPlaybackRuntimeRepository{
 
 	@Override
 	public PlaybackRuntimeDescription findById(String playbackGatewayId) {
-		return delegate.findById(playbackGatewayId).orElseThrow().convertTo();
+		PlaybackRuntimeDescriptionDO obj = delegate.findById(playbackGatewayId).orElse(null);
+		if(Objects.isNull(obj)) {
+			return null;
+		}
+		return obj.convertTo();
 	}
 
 	@Override
 	public void deleteById(String playbackGatewayId) {
-		delegate.deleteById(playbackGatewayId);
+		if(delegate.existsById(playbackGatewayId)) {
+			delegate.deleteById(playbackGatewayId);
+		}
 	}
 
 }
